pyproject.toml
[project]
name = "simple-auth-jwt"
version = "0.2.0"
description = "Plug-and-play JWT authentication SDK for Flask and FastAPI"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name="Omar Rodrigues"}
]

dependencies = [
  "passlib[bcrypt]",
  "python-jose",
]
